feat: creates /investments/buy endpoint that allows the client to buy an asset
creates a record in the investments_history table with the account id, asset id, investment type (buy or sell), purchase price and quantity. Also creates a record in the portfolio table, adding account id, asset id, quantity (assigns the current purchase quantity), symbol and average_price (assigns the current purchase unit price). If there is already a record in the portfolio table with the account_id and asset_id, it just updates the amount in the portfolio, adding the amount purchased.

It is still necessary to create a logic to update the average_price